requirements.txt
openai-whisper
python-dotenv
watchdog
pyyaml
